# File cache optimization