--- a/extension/httpfs/src/include/cached_file_manager.h
+++ b/extension/httpfs/src/include/cached_file_manager.h
@@ -4,6 +4,7 @@
 
 #include "common/case_insensitive_map.h"
 #include "common/file_system/virtual_file_system.h"
+#include "common/types/types.h"
 
 namespace kuzu {
 namespace httpfs {
@@ -29,12 +30,13 @@ public:
 
     ~CachedFileManager();
 
-    common::FileInfo* getCachedFileInfo(HTTPFileInfo* httpFileInfo);
-
-    void destroyCachedFileInfo(const std::string& path);
+    std::unique_ptr<common::FileInfo> getCachedFileInfo(HTTPFileInfo* httpFileInfo,
+        common::transaction_t transactionID);
 
 private:
-    std::string getCachedFilePath(const std::string& originalFileName);
+    std::string getCachedFilePath(const std::string& originalFileName,
+        common::transaction_t transactionID);
+    std::string getCachedDirForTrx(common::transaction_t transactionID);
     void downloadFile(HTTPFileInfo* fileToDownload, common::FileInfo* cacheFileInfo);
 
 private: