--- a/extension/httpfs/src/s3fs.cpp
+++ b/extension/httpfs/src/s3fs.cpp
@@ -56,8 +56,8 @@ S3FileInfo::~S3FileInfo() {
     }
 }
 
-void S3FileInfo::initialize() {
-    HTTPFileInfo::initialize();
+void S3FileInfo::initialize(main::ClientContext* context) {
+    HTTPFileInfo::initialize(context);
     auto s3FS = fileSystem->constPtrCast<S3FileSystem>();
     if ((flags & O_ACCMODE) & O_WRONLY) {
         auto maxNumParts = uploadParams.maxNumPartsPerFile;
@@ -128,7 +128,7 @@ std::unique_ptr<common::FileInfo> S3FileSystem::openFile(const std::string& path
     initCachedFileManager(context);
     auto s3FileInfo =
         std::make_unique<S3FileInfo>(path, this, flags, context, authParams, uploadParams);
-    s3FileInfo->initialize();
+    s3FileInfo->initialize(context);
     return s3FileInfo;
 }