# Description

This PR improves the file cache performance by assuming cache files have no changes inside a transaction. Each query inside a transaction will share the same caching file.

Benchmark result on scanning remote LDBC-10 COMMENT table:
query: `match (c:Comment) return c.*`
Result:
|  DB | attach time | first scan time | scan after first scan time |
|----------|----------|----------|----------|
|   Duckdb  |   0.152 s  |   9.06 s  | 0.146s |
|   Kuzu  |   2.259 s  |   1.676 s | 0.490s |


Load from LDBC10 comment.csv performance:
query: `load from 'comment.csv' return *`
|  DB | compile time | exec time | total time |
|----------|----------|----------|----------|
|   Duckdb  |     |     | 2.38s |
|   Kuzu master |   2.801 s  |   10.35 s | 13.1 s |
|    current branch | 2.772 s  |   2.32 s | 5 s |