@ray6080 2024-05-23
I remember you said this should improve the performance of `LOAD` csv files? Can we see some numbers over that?

Also, where did you clean up the cached files when transactions commit? I don't seem to follow that logic.